Migrate API to FastAPI
Migrated API to FastAPI from Flask to take advantage of its
better application lifecycle management and other features. In
order to do this, the explorer data model was migrated from
dataclasses to pydantic to be compatible with FastAPI. Docker Compose
files were also updated to use the appropriate commands for launching
the FastAPI app with gunicorn and uvicorn.